Fixed type errors, extracted static camera into own files

+++ source/graphics/SceneConfig.h
@@ -2,5 +2,5 @@
 
 #include <cstdint>
 
-typedef int64_t SceneObjectId;
+typedef int32_t SceneObjectId;
 static const SceneObjectId invalidObject = -1;
