+++ source/util/Time.cpp
@@ -5,6 +5,6 @@
 double getTime()
 {
 	auto time = std::chrono::high_resolution_clock::now().time_since_epoch(); 
-	double microsec = std::chrono::duration_cast<std::chrono::microseconds>(time).count();
+	double microsec = (double) std::chrono::duration_cast<std::chrono::microseconds>(time).count();
 	return microsec / 1000000.0;
 }
